m3g: Optimize rendering by using DataBuffers and direct conversion
Image2D (M3G in general) works with RGBA for alpha images whereas
FreeJ2ME's Graphics objects work with ARGB... so to be able to use
DataBuffers for faster rendering, what we need to do is make sure
that the getPixel() method of Image2D returns the pixel value
converted from RGBA to ARGB. Using DataBuffers instead of a
WritableRaster also dropped the need to have the getPixelArr()
method.

Race Driver GRID reports an average of 24-26% higher performance
with these adjustments, due to its unlimited framerate.