[EXPERIMENTAL] PlatformGraphics: Faster drawRGB calls
Trying to use direct dataBuffer manipulation in drawRegion can
introduce performance regressions, whereas all tested cases in
drawRGB have resulted in a performance gain, so we're going with
it as is, but introducing some validation checks.

Also, nokia's drawImage had no reason to issue a call to drawImage2()
while all it really needs is a direct call to gc.drawImage().