[EXPERIMENTAL] Fully rework input handling again.
This time, no dedicated runner to handle inputs, as this caused
issues whenever the input code was in a separate core from the
game logic, breaking games like Ratatouille and Heroes Lore: Wind
of Soltia. Instead, run it on the main FreeJ2ME thread and run
checks to see if the input code has to do something or not.

Given how far-reaching this is, i won't be surprised if it breaks
other stuff. So i'll be marking it as experimental.

Fixes #7 and #17, though i didn't test all versions of Ratatouille