m3g: Graphics3D: Support frustum and backface culling.
NOTE: There is a hack that disables backface culling for triangles
that have been clipped. For some reason, their normals aren't in
line with the original triangle. In cases where the original was
facing the screen for example, the new clipped one(s) would end up
facing away from the camera and so on. No idea why that is the
case.