RecordStore: Change where the record store trackers are updated
Let's do it right at the end of the RecordStore constructor, since
openRecordStore will always issue a call to it anyway, and doesn't
need to update those trackers manually in case of issues. Fixes #55
and probably quite a few others when creating a new record that
doesn't exist (and has the "createIfNecessary" flag set as false)...

I'm a bit impressed that this didn't cause as much trouble as it
could have.