Treewide: Assorted fixes, and making sure Boombakas Pro works
So we finally found a case where the handling null images has to be
configurable. Boombakas Pro REQUIRES null images to throw an
Exception in order to be playable, whereas House M.D. breaks IF
an exception is thrown because it's buggy and doesn't handle
NullPointerExceptions. So this basically means we now have the
very first compatibility setting in FreeJ2ME.

Also, Boombakas Pro checks for the "supports.mixing" sys property
at boot and crashes if it is not present... which meant it was
a good time to work around some sys properties that were missing
beyond this one.

Beyond that, this game also loads media files from a DataSource,
so that had its implementation finished and tested.

Aside from graphics flickering, the game seems to be perfectly
playable now. Related to #59.

Other than this, there were some minor fixes and cleanups that
wouldn't warrant their own commit, like the DirectGraphics pixel
format matching KEmulator and J2ME Loader, canvas invalidating
the screen contents properly on a jar fullscreen change, etc.