lcdui: Also serialize inputs and UI routines
Now we don't need a bogus "delay" in input polling, as Ratatouille
and Heroes Lore: Wind of Soltia work properly with this improved
serialization.

Canvas' serviceRepaints is probably not accurate yet, as it just
forces a screen redraw to happen without much care for anything
else, but it'll do for now.

Fixes #60 and also resolves the last comment in #38.