Graphics: Improve clip region handling
Previously FreeJ2ME was not consistent in regards to the clip
region that it has set, and what it would send to jars when they
requested data about the clip region and position... it also wasn't
following the MIDP spec too closely. These new changes improve it
by cleaning up the older clip code, and using a single, consistent
Rectangle object to keep track of the clip bounds.

Fixes #62 and probably many others that rely on clip region
information.