Introduce a compatibility setting for Fantasy Zone 128x128
Its requirement of issuing a clipRect() call instead of setClip()
whenever the graphics object is reset for screen rendering breaks
a lot of other games, so turn this into a compat setting as i can't
think of another game that needs this. Fixes #64, and is also related
to #10 (of course it is, that issue is massive).