M3G: Implement animation track checks and updates.
This doesn't seem to be used all that often, but it's over-arching
enough that it should be included soon into the API's support.

Also prepare Graphics3D to draw stuff through render(World) and
render(Node, Transform) beyond just the direct call to
render(VertexBuffer, ...). Render(World) seems to work properly,
and render(Node, Transform), which does the bulk of the work, is
still missing some critical parts, but this is already a massive
commit even without them.