Canvas: Drop old compatibility hack for recursive repaints.
This is approaching a point of no longer having benefits, so might
as well remove it from the code. Only leaving the notification as
a debug print.

Fixes Rainbow Six Urban Chaos and allows Crash Bandicoot 1 back
into game (if limited to 15, which is what the game kinda runs
at). Related to #37 and #75.