LCDUI: Introduce a paint queue and rework Display.setCurrent()
Quite a few changes on this one, so i'm sure there'll be tons of
regressions, particularly in horribly programmed jars like Heroes
Lore: Wind of Soltia and Ratatouille.

For starters, now repaints and lcdui updates are queued in their
own event queue (might share space with inputs at some point), and
now serviceRepaints() does what it is meant to do, which is forcing
all pending repaints to occur by blocking when called. This is more
in line with the documentation, and how J2ME-Loader does this.

setCurrent() requests do not take effect immediately anymore (also
more in line with the documentation), instead it gets saved into
a Runnable that is then read by the painter thread to see if a
change of current displayable has to be made. This has the benefit
of not needing to slot parts of setCurrent() into the callSerially
queue like it was previously done.

Furthermore, recursive checks for repaint and invalidation are now
gone, they were doing more harm than good at this stage.