Partially Restore input handling to the older system
It still maintains the improvements done since it was the default,
just the "processInputs()" method and the related polling mode were
removed in favor of the more direct command issuing. I can already
see this causing regressions on the SDL frontend, but given how
little i'm focusing on it at the moment, i don't see it as much of
a loss, where this will very likely reintroduce the issue of no
inputs being read when the game or app doesn't draw new frames until
a new input is registered (which basically means a softlock).

But other than that... this commit alongside the one immediately
prior to it should fix a BUNCH of recently discovered regressions,
like #75, #71, #73 and to some extent (although untested), #66 at
the same time maintaining compatibility with games that were fixed
by the newer code, for example, Racing Fever 1 & 2.