LCDUI: Finally implement TextField and TextBox input handling
TextBox is completely untested, but given how much it shares with
TextField, which has been tested at some basic level and works, it
should also be fine.

Multiple Charsets are also supported, although not all of them are
fully featured (this is already a massive commit, and i'm not versed
in much else other than Latin and some Japanese), it falls on the
application to request the correct charset based on the Oracle MIDP
documentation. Line breaks are also supported.

Quick rundown of the navigation for those:
Arrow Up: Move to the next character of the charset. (wraps)
Arrow Down: Move to the previous character of the charset. (wraps)
Arrow Left: Move the caret to the previous position, allowing edits
Arrow Right: Move the caret to the next position if possible
Fire Key: Add the selected char into the current caret position
Star Key: Remove the character at the current caret position
Pound Key: Add a space at the current caret position.

Related to #69, further testing is VERY welcome.