LCDUI: Blind-implement more CustomItem parts, add a check in Form
In some cases, it was possible for the form to contain more items
than item bounds, so let's guard against such cases when rendering.

CustomItem is VERY rare so jbenchmark2 is the only thing i've seen
create a CustomItem ever, and even that does little more than just
create the thing, so testing it will be a bit complicated... which
is why its WARNING log was changed to point out it is untested upon
creation.

That said, i don't think there's much else to actually implement in
LCDUI, only refine.