Stub the Micro Edition Content Handler API
This one doesn't seem to be a MIDP standard... however, Where's My
Water requires this to boot. It still complains about being unable
to open upp the In-App Payment library (because this is a stub), but
the bundled levels can be played just fine aside from MPEGPlayer being
a bit spotty.

Siemens' io.Connection class had its constructor made public in
order to use it as a Connection replacement for this API's Invocation
class, just so we can keep this stub 1:1 with the specification.