RecordStore: Properly keep track of record IDs
Previously this was not done at all, and recordIds were erroneously
based on the record's position on the "records" vector, as well as
the vector's size. This commit introduces a second vector to keep
track of each record's id, to further align this implementation to
the documentation, which specifies that record IDs should never
be reused (and they would be with the prior setup, especially when
loaded from a file).

Of course, this means that this new vector also has to be saved
in the rms file, and as such, this is a breaking change. To minimize
issues, those record ids are saved AFTER lastModified, so that instead
of breaking absolutely every save, this only breaks saves made/opened
before commit 661761d1fe4919a7b0a103acff3b50bfd488dc02 and that were
NEVER once re-saved between that commit and this one, which hopefully
isn't doesn't affect many out there...

Related to #59 and #77.