PlatformGraphics + Libretro + SDL: More rendering optimizations
We don't need to allocate the dataBuffer arrays all the time we're
going to read/write them, only when the underlying image actually
changes. This results in a performance increase of about 10-15% in
AWT depending on the scene, while Libretro averages double that
because its screen rendering also uses dataBuffers (it is now faster
than AWT), and SDL also received a similar boost to Libretro.