MIDletLoader: Rework how the manifest is parsed again
Rethinking everything, this previous implementation wouldn't work
well at all (icons inside a folder for example), let's go for a
simpler approach that only splits based on the "/" char if really
needed, and ONLY at the first occurrence. If the manifest is lacking
a comma between the icon path and the main class name, i don't think
it'll be possible to fix it here.