PlatformGraphics: Add Null check in setFont() and cast short pixels
Previously we didn't check if the received font was null,
Castlevania 1's Spanish 176x208 version sends a null font and broke
caused FreeJ2ME to crash. With that out of the way, the game then
sends short type pixels to be drawn in the type reported by
DirectGraphics' getNativePixelFormat() function, and unlike karma
games, it doesn't care if the type is valid or not to then send
pixels in the format they are natively in so that FreeJ2ME can convert.

As a result, another crash due to exceptions being rightfully thrown
because "0" is not a valid pixel enum type. So now FreeJ2ME will
forcibly decode such cases as if they were TYPE_USHORT_4444_ARGB,
as this format is the closest to the native screen's. Game now works
fine.