Frontends: Minor input adjustments so that LCDUI commands work
Now everyone should be casting properly to the same input values
that libretro does (which is the simplest as far as number of
keys are concerned). Fixes probably everything that relied on
Canvas getGameAction to process commands in AWT and SDL.

Also reduce the size of the pressedKeys array, we no longer need
all those keys... or so i think.