PlatformPlayer: Substantial improvements to MPEGPlayer
Since "Where's my Water?" is a great candidate for testing, and
exposed many issues with the currently MPEGPlayer implementation,
it was time to sort it out. Now it manages playback states far
better and doesn't need to have null checks all over the place.

It also retrieves and sets the media positioning far better than
before, although it's still not perfect (and well, the documentation
doesn't state that the positioning should be perfectly precise).