sprintpcs: Clip: Simplify and optimize locator loading
We don't need to constantly get data from the jar file. Send it
as a byte array to the stream variable and keep it there in case
the player needs to reload it. Way faster and less prone to stuttering,
while also making getPlayer() simpler.

Mobile also has a getMIDletResourceAsByteArray() method now, to
simplify getting things from the MIDlet.