javax: Separate and stub/implement most relevant Connection classes
This means that SMS message handling and http should now be handled
in some capacity (enough for games that truly use messaging and
MessageConnection not to crash or fail at least). Actual data input
and output remains stubbed, although a simple bypass for vServ's
wrapper has also been grabbed from KEmulator in order to allow
apps that use this to work.

Related to #81.