LCDUI: Rework command issuing (again) and simplify keyState logic
Let MobilePlatform handle command issuing into the Displayable
objects, as it is the class where every single input, no matter
the API or vendor, goes through, so we're guaranteed to be able
to issue LCDUI commands from there in cases where they're needed.

Fixes a bunch of stuff as far as commands and input handling go,
for example Impossible Mission, many SoftBank games and Football
for Nokia 3410.