More cosmetic improvements to MIDlet suspend and resume.
I keep forgetting that there's stupid jars that refuse to update their
screen data when they're brought back into active state for some reason,
despite the documentation clearly stating that the device screen's
contents might not be consistent with what the application last drew.

The workaround is AWT-specific since Libretro has no way of painting
an overlay on the screen that does not fall into the same issues AWT
was having prior to this commit, same for SDL AFAIK, as both render
directly from the mobile display. At least with libretro the overlay
isn't ever going to be seen anyway, as the frontend menu is what will
get the user's attention.

Also also related to #83.