EXPERIMENTAL: Introduce an AWT media player and separate OTT decoder
It was about time, since this opens up FreeJ2ME's decoders for SMAF
and OTT/OTA to work on data outside of streams bundled in jars. This
might also help any future decoders for AMR, etc.

As a bonus, this commit also adds drag and drop support to the AWT
GUI and further improves the indicator when the MIDlet is paused.

Closes #85
Closes #84