DoJa: Prepare to actually load scratchpad files
We can't just load it normally like any other stream, for starters,
the scratchpad isn't placed inside the jar, furthermore, it can
be read from, and written to, which demands a more robust structure
to hold its data. That's where ScratchPadConnection comes in,
initially as little more than a beefed-up InputConnection, but which
will be extended with more functionality once the Scratchpad format
is worked out.