EXPERIMENTAL: Rework how the Graphics objects extend from each other
PlatformGraphics is now the main abstract class that shall implement
ALL methods from spec and vendor Graphics classes, be it lcdui's
Graphics, DoJa's Graphics, MMPP's GraphicsX, etc. In turn, those
classes now extend it directly with little more than a constructor
for them. This allows the base Graphics subsystem to be reused in
all vendor implementations with little to no adjustments or casts.

This still has the potential to break some jars simply due to how
broad of a range it has (it IS a massive change to the Graphics
subsystem after all), so it's marked as an experimental change that
may be partially reverted in the future.