EXPERIMENTAL: MIDletLoader: Override jar's System.gc() calls
Because we're not working with minuscule heaps like those devices,
and we're running a much newer VM, let's assume the JVM itself knows
best about when and how to garbage collect, and take this out of the
MIDlet's hands... Especially since some jars like to abuse gc calls
in-between frames,  which can cause serious performance degradation.

This gc-happy behavior is probably a method to keep heap usage at a
minimum on those devices, but here, constantly forcing GC to run every
frame is way too costly on CPU and we have memory to spare.