M3G: Rework how triangle clipping is done.
Now clipping will no longer break down and re-triangulate triangles
at the edges of the screen, it either maintains the original triangle
if it's partially visible, or removes it from the draw call if it's
fully offscreen (we just skip drawing pixel of the triangle that are
offscreen). This is closer to how OpenGL works, and in turn fixes
instances where the clipped triangles would result in incorrect
vertex colors, texture distortion and a potentially bigger list of
triangles to render.

So to sum it all up, with this change:

1 - M3G renderer is slightly faster
2 - There's less texture distortion (still present, but not as much)
3 - Texture blending is improved, but still not accurate to OpenGL

As notable fixes, the background in Brick Breaker Revolution 2 3D
is now a circular shape as it should be instead of a squircle,
and Tornado Mania 3D's terrain colors are now consistent at the
edges of the screen.