EXPERIMENTAL: Display: Improve paintQueue and callSerially
Give serial calls a small breather in the form of a small Thread.sleep
of 1 ms between every processing. Some games really spam this queue to
the point of stalling even the render thread, like Heroes Lore: Wind
of Soltia does for example. Furthermore, we don't actually need to
clear the paintQueue when a new displayable is made current, and
serviceRepaints should issue a notify call after all pending
repaints are serviced.

Fixes Ratatouille, and Heroes Lore: Wind of Soltia... though they
really shouldn't be relying this kind of behavior.