Display: Rework how the serial calls work in relation to repaints
According to the documentation, serial calls should happen after
each pending paint cycle (assuming there's any, if not we can just go
and process serial calls one after another).

So the plan now is to not even have a separate thread for serial calls
anymore, instead, they're queued as normal, but processed by the
paint thread, AFTER the current paint cycle is completed, and in a
non-blocking manner.