RecordStore: Improvements to RecordEnumeration, use RecordListener
Related to games brought up by issue #93. Fixes both of them by
improving RecordEnumerator's enumeration logic and record retrieval,
as well as implementing the usage of RecordListeners for it, instead
of forcing rebuild() calls every time a method from Enumerator was
called (which was incorrect as rebuilding should only happen if there
was an actual change to the RecordStore).