EXPERIMENTAL: Canvas: Set clip to draw area on repaint.
This is what J2ME-Loader does, and seems to help some games like
Fantasy Zone's 128x128 version (no longer needs the hack to render
something to the screen), without any regressions so far.

Also, Fantasy Zone's compatibility setting had a change in
functionality, now it translates the graphics object to the origin
on every draw, otherwise FZone constantly translates the image away
from the visible area.

This reeks of a bug hidden in the graphics rendering code, but i've
no idea what it is or how to properly fix it.