EXPERIMENTAL: Display only retrieved from Mobile, setCurrent fixes
Instead of making a mess of lcdui.Display by having it spread
between Mobile and MIDlet, have it exclusively set up on Mobile,
and created alongside MobilePlatform to prevent issues where
the Display is not setup before the app calls upon it for anything.

Beyond that, some versions of Rayman 3 queue up repaints right on
a setCurrent call, which could cause issues with the method's usage
of notifySetCurrent to render the displayable, so what is now done,
is to service any pending repaints BEFORE rendering the new
displayable.