Replace the ignoreGCCalls compat setting for ImmediateRepaints
Some games and apps might deadlock by queuing up paint events that
acquire a lock required by something else, and expecting to have
this lock available BEFORE the paint call is processed, so for
these cases, this new setting might help.

We don't need to have a setting to ignore explicit gc cals anymore
since the instances where it caused issues have been fixed, and it's
quite likely that it won't be the actual source of any new issues.