RecordStore: More fixups to RecordEnumeration.
So the last round of RMS improvements fixed two chinese jars that
used RecordEnumeration but expected the index to be at the end of
the record array... BreakQuest Mobile actually expects the index
to be zero. So make sure both cases are handled by the RMS subsystem
without requiring a compatibility setting.