PlatformGraphics: Improve how translations are handled on reset()
Some jars like Sonic 1 for Panasonic VS3 and DOOM RPG for the
Motorola V551 translate the drawing area before drawing anything to
it, i assumed previously that we shouldn't translate it all the way
back to 0,0 on every graphics reset because it broke those jars, but
then others like Fantasy Zone's 128x128 version will keep translating
the canvas away from the screen if the translation isn't reset.

The compromise here is to check whether the jar is calling its first
reset, and if so, save the current translation at that point and then
always reset the translation to that going forward. This way, both
cases are fixed, and Fantasy Zone no longer requires its "translate
to origin" compatibility setting. It's likely that this also fixes
many other instances of misalignment.