Canvas: Introduce a mechanism to get out of serviceRepaints lockups
Some versions of Rayman 3 can paint on the same thread that it calls
serviceRepaints on. In those cases, the repaints would never be
serviced and it would block infinitely. Introduce a check to see
if serviceRepaints has been blocking for more than 100ms without
interruption, and force a repaint to happen if so.