EXPERIMENTAL: PlatformPlayer: Deallocate + close players in thread
For wav, the bulk of the stutters and short freezes seem to come
from deallocating and closing the player, so doing these in a
thread allows those methods to return instantly and prevent said
performance degradations. Since it's relatively safe, this change
was made to all available players to improve responsiveness on jars
that constantly close and open new players.

With close() this should be completely safe since a closed player
is never meant to be reused again, now deallocate() on the other
hand requires some testing to make sure it's also safe, although
i doubt jars would deallocate a player just to instantly prepare
it for playback again.