Massive overhaul to how RecordStore data is handled, and saved
Probably the biggest rework on RecordStore so far, tHe biggest factor
here is that now saves are cross-compatible with SquirrelJME!

Although the saved data layout is wildly different from the prior
implementation, FreeJ2ME-Plus is able to convert those older saves
(even original FreeJ2ME's saves) to this new format without issue,
so no saves should be lost in the transition.

But other than that, tag handling has been implemented, access to
records from other vendors and suites as well, the savedata is both
more readable and robust, and much, much more.

Has the potential to break some stuff, but this is not something
i aim to revert in any way thus it won't be marked as EXPERIMENTAL.

Besides, this is also the gateway to handling DoJa writes just like
SquirrelJME does through RMS, which will allow cross-save support
for DoJa as well.