DoJa: Canvas: Rework how inputs are sent to processEvents()
Bomberman Jetters is a jar that doesn't override processEvents at
all, so it was a good way of checking whether inputs were being
sent properly... and a few changes were made.

Also changed all update*KeyStates functions signatures to receive
a boolean indicating whether the key is pressed or not instead of
an integer.