EXPERIMENTAL: PlatformPlayer: Multiple improvements to MIDI playback
Now even the Sequencer is shared between midi players, which is to
say, pretty much everything is shared now. Memory footprint should
be even lower when multiple MIDI sequences are involved, and MIDI
in general should have a lower runtime overhead due to this.

Probably some parts of state management will need more work to
better differentiate between sequences to be played... especially when
they're meant to be played concurrently like BGM and SFX, where we
should actually pause the currently running BGM, run the SFX and then
resuming playback of the BGM, which is completely untested, but so
far, this helps volume changes work better, and the MIDI system is far
smoother even on very low-end devices.

SMAF also handles mixed playback a tiny bit better by checking if
the PCM index it's going to play is actually populated before accessing.

Tested on an AMD-C60, and stutters involving MIDI player creation are
nearly gone in Java 8, and basically a no-factor in Java 11 and newer.