Libretro: Further improvements to how windows loads FreeJ2ME's jar
Go back to using sprintf() calls to concatenate the cmd strings,
except that this time we'll be using snprintf to prevent warnings
and introduce better protection against overflows.

Previously the usage of strncat and strncpy could cause issues for
special chars in some encoding formats since the size parameter in
them reflect the size of the char buffer, and not the amount of
characters in the string inside them. Hopefully this also improves
the last issues brought up in #108, otherwise the fix for those would
require some serious rework on how this loading procedure is done to
then support multiple different encoding formats.