AWT: Support reading external input events, also rework features
Paused is now governed by Mobile instead of MobilePlatform, and
having an external file for external events allows FreeJ2ME to
work with gamepad APIs that run outside of Java, so no need to rely
on JInput or any other library that only works on the usual Windows,
Linux and Mac platforms!

But beyond that, what this also means is that now the web version of
FreeJ2ME is ready to go! Coming really soon to the webpage...