LCDUI: Rework how events are serialized again...
Alright, i think the general idea for this if finally settled...
This structuring should be the one to use going forward, with
improvements done over time to fix any rough edges. This follows
the MIDP specification as closely as possible without harming
compatibility with problematic apps. For instance:

- Inputs are queued in relation to the serial events, but added in
a secondary thread since the main thread could be overloaded by
multiple inputs being added at the same time (Konami's Rush and Attack
being a prime example)

- Every method related to serial events and paint calls is synchronized
on top of the same event queue, even if the events themselves are placed
into / polled from another Runnable or queue. This is so that the
Event Processing Thread is always the one to dispatch those events, and
they retain their serial processing nature in some shape or form.

- Queued inputs are always fully processed before the paint event is
completed, this is so that jars like Ratatouille don't get stuck on the
menu by having the event thread be blocked by any lengthy processing before
it can do anything else.

- SetCurrent requests are always honored AFTER inputs, pending paint events
and at least one serial call are processed, so that the chances of an event
that's meant to occur on the previous displayable being missed are lower,
most notable in jars like Harry Potter - Find Scabbers and Racing Fever 2.

- Canvas repaints and serviceRepaints, while synchronized on the same
serializedEvents queue (like everything else), actually post their events
separately, thus allowing serviceRepaints to make sure it's only blocking
in order to process any pending paint events, with the actual serial queue
being unchanged by it so that events don't get misplaced or processed when
they shouldn't. The wait timer for paint events to be processed by the event
thread, however, remains. This is so that jars that rely on serviceRepaints
timing like Gradius NEO (MIDP) and some Rayman 3 versions work as expected

- The event thread can now wait for as long as there are no events to process,
as since all events are synchronized on the serializedEvents queue, this allows
any event additions to notify the event thread to resume execution. This not only
reduces CPU usage in situations where the queue is emptied for some time, but is
also bound to improve event timing since there's no longer a constant sleep(1) call
forcing the thread to sleep and wake up when the queue is empty. Instead, the thread
wakes up as soon as the queue is non-empty, without an arbitrary sleep value.

- SetCurrent now handles if the displayable's notifySetCurrent call should be
explicitly queued or not based on the displayable's subtype. This is because
Canvas already queues all its repaints by default, so having the Displayable
superclass always queue its notification calls for LCDUI components (should)
and Canvas (shouldn't, it will be queuing two events instead of one, and only
the last would be an actual repaint) could introduce issues.