lcdui: Simplify the event synchronization logic.
Now Canvas shouldn't need to synchronize on Display's event queue,
allowing it to be private to the Display class... also turn paint
and setCurrent events into AtomicReferences, like DoJa's Display is
currently doing. Beyond that, take the setting of pendingRepaints to
false out of the repaint call, do it right after it returns.

We should also check if the current input event is null before
processing, as it's a very active queue depending on the application.