EXPERIMENTAL: Target Java 1.6 as the lowest VM version
This required quite a few changes in FreeJ2ME's classes, although
nothing too extensive. Still, since Java 8 will be used to build it
targeting Java 1.6, bugs could still arise when running FreeJ2ME on
it, and Java 8 remains the baseline for actual support and bug
reports.

As a bonus, M3G's required some changes to how triangles were processed
in order to comply with java versions older than 8, and in doing so,
i found out a few ways in which to optimize this section of the render
path, especially concerning memory pressure, which means M3G should
not only be faster now, but also have much more consistent frametimes
than before.

With all that said, this commit might cause regressions.