DoJa + LCDUI: Fade the command bar.
Some jars draw into the area occupied by the command bar, so fade
it after 3 seconds so that the user can see what's behind it. If
the command bar is not visible, pressing any of the soft keys, or
having any command added/changed/removed by the app will make it
visible again. If the bar is visible, pressing any of the soft keys
will issue the command tied to it.

If the jar doesn't draw into the command bar's area, it still fades
by making the text gray.