SMAFDecoder: Support volume changes for PCM in mixed playback
This is done in a rather clever way: PCM sample volume at any given
time is set based on the note velocity. Mapped from -30dB to 0dB with
the velocity range going from 0 to 100 (higher values are clamped to 100)
makes it sound reasonably close to Yamaha's MidRadio.

Panning is also done in some samples, but i don't think Java Sound API
has any way of simulating it... though it appears to be done in SMAF
through the pitch bend value.

Related to #115.