WAVImaADPCMDecoder: More improvements and optimizations to decoding
Malformed blocks no longer seem to be a thing after using the
sampleDataLength field straight from the WAV header itself. So not
only this simplified the decoder quite a bit, it also makes it faster
and more memory efficient by dropping the explicit checks and no longer
requiring partial array copies at the end of the decode step.