WAVTools: Further improve and optimize upsampling method
We can just pass the correct amount of samples to decode into the
upsample method in order not to touch padding or garbage data. WAV
headers have a field just for that, and any ADPCM decoded data should
be unaffected by this.

The end result is that the upsample method is now smaller, a bit faster,
and no longer introduces crackling on streams whose amount of valid
samples doesn't match its data array's size.

As a bonus, a few "final" keywords were introduced on some pertinent
methods and classes... we won't be extending those ever.

Fixes #118