DirectGraphics: Un-translate clip region before drawing images
Fantasy Zone's 176x208 version requires the clip position to not
be translated when drawing images through DirectGraphics' drawImage()
method. It uses both DirectGraphics and lcdui.Graphics for rendering
at the same time (directGraphics for transformed images, lcdui for
normal images), hence why the clip is set back to the original position
at the end of the call.

Does not seem to cause regressions on other jars that use DirectGraphics
like Sonic 2 Crash.

Related to #21