Display: Isolate each major block of setCurrent in try-catches
Castlevania 1's High-End Chinese version has a bug in its code
that makes it read from a null array when its canvas is set to be
shown on boot. Isolating the showNotify() block from the rest of
setCurrent's code allows it to soldier on perfectly fine after
catching the NPE thrown by the bug.