Display: Changes to setCurrent logic, and event processing
Marcel Desailly Soccer is now playable, because it is one of the
rare jars that have issues with setCurrent forcing a canvas repaint
when making it current.

What happens now is that FreeJ2ME-Plus waits for a max of 200ms for
the jar to set up everything it needs to call a canvas repaint, and
even call the repaint itself, removing the need for setCurrent to
forcibly do it in such cases.

As a consequence, it seems that paintEvents can now be safely processed
always before any serialCalls, like the MIDP documentation states.

Related to #37