MIDletLoader: Tweaks to platform properties.
microedition.profiles is now reported as "MIDP-2.0", some versions
of Ratatouille won't even boot without it. Also make microedition.encoding
report the same as the current session's file.encoding property.

A prop for mms was also added but i doubt it's useful for much, and
it's just a stub too.