Siemens API: Bunch of fixups to mp.game classes.
Spite and TiledBackground classes are now far more robust, with
validation checks and allowing the mask image to be null, as it is
not required to be present.

Also, ExtendedImage's setPixels was reworked to be more accurate by
testing against multiple jars that previously had issues with it
like Shooter, Rayman Golf, Kool, and others.