PlatformPlayer: Major improvements to MIDI/SMAF playback.
Now multiple sequenced sources can play concurrently (up to 4, as
that is now the number of exclusive synthesizers), the core logic
was also reworked to handle that, and one of the synthesizers is
shared with Manager's playTone method to save on memory, as synths
use quite a bit of memory by themselves.

4 concurrent MIDI/SMAF sequences should be enough for just about
anything, and KDDI notes in their documentation that up to 4
concurrent streams can play at any given time.

Furthermore, those memory-heavy operations are now done immediately
prior to starting the application, minimizing any stutters that could
happen whenever they started interacting with either Manager or
PlatformPlayer classes.

Related to #115