ContentHandler API: Implement a few parts of it.
This was mostly in the hopes that Nokia's In-App Purchase API (not
to confuse with Nokia In-App Payment API) could have purchases bypassed,
but it's a much more complex API and looking at games like Angry Birds
and Where's My Water? show that they look for specific server responses
and any bypasses would have to be done per-app. Not really worth the
hassle, but i didn't want to ditch the work done on CHAPI either.