PlatformPlayer: Some improvements to sequence players.
Volume changes are handled a little better now, and there's no
longer any need to re-add a MetaEventListener to the sequencer at
every start, as we're not reusing the same MetaEventListener for
multiple players anymore.

Also took the time to reorganize some code block around, mainly
related to synthesizer handling, so that The logic doesn't have
to jump around PlatformPlayer and Manager as often.