MIDletLoader: Improve overseas vendor checks and encoding setup.
Checking for "KDDIP" on the Microedition Profile string wasn't a
very good way of detecting KDDI apps, as not all of them set this
property. It would be possible to do that just by checking whether
the file has the kjx extension or not, but that would break extracted
KDDI apps, and would mean yet another vendor having checks all over
FreeJ2ME-Plus like DoJa has (but DoJa really needs them).

So the alternative is to move the bulk of those checks into
MIDletLoader and base them off of vendor-specific classes. If an
app loads any KDDI class (or J_Phone one), the KDDI flag is set to
true, and the same is done for DoJa there, even if not needed.

This makes such checks much easier to expand upon at a later date.