Mobile: Rework how text encodings are handled.
"UTF-8" is no longer a thing, and FreeJ2ME-Plus will restart itself
to ISO_8859_1 at boot if it finds itself using an encoding format
that's not one of the supported ones (those being ISO_8859_1, Shift_JIS
and EUC_KR at the moment). This fixes the issue where jars that use
ISO_8859_1 would end up using the System's default encoding, while
maintaining the functionality of switching to EUC_KR or Shift_JIS
based on heuristics. EUC_KR is changed based on the presence of a
MSD file for SKT, and Shift_JIS is changed based on DoJa properties,
or the loading of classes known to be tied to a Japanese Vendor
implementation.