MLDDecoder + SMAFDecoder: More big improvements to both
Now MLD is no longer an experimental piece of code, it works reasonably
well for sequenced data and even supports PCM decoding, although the
results vary a lot in quality (and they seem to play too loud compared
to SMAF).

SMAF on the other hand, mostly received optimizations and further
improvements to Handy Phone and Softbank formats, but the main change
here is that it now supports Huffman Decoding.

Closes #114 and also #115