Fixing some regressions all around.
Custom fonts and sizes now work again, the canvas command bar is
rendered straight to the front buffer so that the application is
the only one fully in control of its graphics context's state,
the limitFPS logic was moved away from canvas objects and is
now handled entirely by MobilePlatform's flushGraphics again as that
makes it easier to maintain as the number of rendering methods increases.

Furthermore, recent optimizations to startup and media allowed tightening
the interval setCurrent waits for the application to automatically render
the canvas, and the WAV upsampler now handles cases where the header's
sampleLength field is incorrect in regards to the actual data size.

Closes #131